switched to FP64 for scalar flux vectors. This greatly improves testing reproducibility and allows for elimination of tolerant test harness variant.

+++ src/random_ray/flat_source_domain.cpp
@@ -114,7 +114,7 @@ void FlatSourceDomain::batch_reset()
 {
   // Reset scalar fluxes, iteration volume tallies, and region hit flags to
   // zero
-  parallel_fill<float>(scalar_flux_new_, 0.0f);
+  parallel_fill<double>(scalar_flux_new_, 0.0);
   parallel_fill<double>(volume_, 0.0);
 }
